Guarantees that the handle will always move after a click
You had to click exactly around the right area to move the handle. This makes it so that each click moves the handle to the next or previous value at least (unless it was already going to move to a further value, or if it was already on the first or last value).